Support for images without src attribute. Real src stored in original attribute.
Works even with Safari. Also make site even faster because even when src attribute is removed with JavaScript some extra images will be loaded because of latency. This is not drop in replacement though. You need to modify your HTML for it to work. This is of course optional feature.

Modified patch by Jeremy Pollock <jpollock_at_brownshoe.com>