Fixing Issue #68
Added a error_callback property that is exposed.
This will be triggered upon error loading an image, and will pass the error data received.
I used the `jQuery.noop()` method so we won't have to add ugly function checks before calling this. I think this is a more elegant solution.